refactor: split tokenizer discovery submodules
